finish with Wall Werror Wextra 😵

+++ libft/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: rabougue <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2015/11/25 21:47:55 by rabougue          #+#    #+#             */
-/*   Updated: 2016/05/04 09:17:48 by rabougue         ###   ########.fr       */
+/*   Updated: 2016/08/10 17:23:15 by rabougue         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "./includes/libft.h"
 
-size_t	ft_strlen(const char *s)
+int	ft_strlen(const char *s)
 {
-	size_t i;
+	int	i;
 
 	i = 0;
 	while (s[i] != '\0')
